Use different screen resolutions in cross-browser screenshot tests

+++ tests/jest-pure/screenshots.test.ts
@@ -1,32 +1,38 @@
-import { Browser, chromium, firefox, Page, webkit } from 'playwright';
-import { launchBrowser, closeBrowser } from './utils';
+import { Browser, chromium, firefox, Page, ViewportSize, webkit } from 'playwright';
+import { launchBrowser, closeBrowser, viewportSizes } from './utils';
 import { PNG, PNGWithMetadata } from 'pngjs';
 import pixelmatch = require('pixelmatch');
 import fs = require('fs');
 
+jest.setTimeout(30 * 1000);
+
 const devBrowser = chromium;
 const compareBrowsers = [firefox, webkit];
-const url = 'https://www.bouvet.no/kurs';
+const url = 'https://www.bouvet.no/investor';
 
-async function setViewport(page: Page) {
-  await page.setViewportSize({
-    width: 1200,
-    height: 2000,
-  });
-  // await page.waitForTimeout(3000);
+const resolutionString = (viewportSize: ViewportSize) => `${viewportSize.width}x${viewportSize.height}`;
+
+async function takeScreenshot(page: Page, viewportSize: ViewportSize, browserName: string) {
+  await page.setViewportSize(viewportSize);
+  return PNG.sync.read(
+    await page.screenshot({
+      path: `${screenshotRootPath}/${resolutionString(viewportSize)}/${browserName}.png`,
+    })
+  );
 }
 
 const screenshotRootPath = `${__dirname}/screenshots`;
-let benchMarkScreenshot: PNGWithMetadata;
+let benchMarkScreenshots = [] as PNGWithMetadata[];
 beforeAll(async () => {
-  try {
+  if (fs.existsSync(screenshotRootPath)) {
     fs.rmdirSync(screenshotRootPath, { recursive: true });
-  } catch (err) {}
+  }
+
   const { launchedBrowser, page } = await launchBrowser(url, devBrowser);
-  await setViewport(page);
-  benchMarkScreenshot = PNG.sync.read(
-    await page.screenshot({ path: `${screenshotRootPath}/${devBrowser.name()}.png` })
-  );
+  for (const viewportSize of viewportSizes) {
+    const screenshot = await takeScreenshot(page, viewportSize, devBrowser.name());
+    benchMarkScreenshots.push(screenshot);
+  }
   await closeBrowser(launchedBrowser);
 });
 
@@ -34,8 +40,7 @@ compareBrowsers.forEach((compareBrowser) => {
   let launchedBrowser: Browser;
   let page: Page;
   beforeAll(async () => {
-    ({ launchedBrowser, page } = await launchBrowser(url, devBrowser));
-    await setViewport(page);
+    ({ launchedBrowser, page } = await launchBrowser(url, compareBrowser));
   });
 
   afterAll(async () => {
@@ -43,17 +48,25 @@ compareBrowsers.forEach((compareBrowser) => {
   });
 
   describe(`${compareBrowser.name()}`, () => {
-    it(`should match ${devBrowser.name()}`, async () => {
-      const screenshot = PNG.sync.read(
-        await page.screenshot({ path: `${screenshotRootPath}/${compareBrowser.name()}.png` })
-      );
-      const { width, height } = screenshot;
-      const diff = new PNG({ width, height });
-      const matches = pixelmatch(benchMarkScreenshot.data, screenshot.data, diff.data, width, height) === 0;
-      if (!matches) {
-        fs.writeFileSync(`${screenshotRootPath}/diff-${compareBrowser.name()}.png`, PNG.sync.write(diff));
-      }
-      expect(matches).toBe(true);
+    viewportSizes.forEach((viewportSize, viewportSizeIndex) => {
+      it(`should match ${compareBrowser.name()} for resolution ${resolutionString(viewportSize)}`, async () => {
+        const screenshot = await takeScreenshot(page, viewportSize, compareBrowser.name());
+
+        const { width, height } = screenshot;
+        const diff = new PNG({ width, height });
+
+        const matches =
+          pixelmatch(benchMarkScreenshots[viewportSizeIndex].data, screenshot.data, diff.data, width, height, {
+            threshold: 0.8,
+          }) === 0;
+        if (!matches) {
+          fs.writeFileSync(
+            `${screenshotRootPath}/${resolutionString(viewportSize)}/diff-${compareBrowser.name()}.png`,
+            PNG.sync.write(diff)
+          );
+        }
+        expect(matches).toBe(true);
+      });
     });
   });
 });
